+++ libs/common/src/lib/dtos/update-account.dto.ts
@@ -4,7 +4,6 @@ import { Transform, TransformFnParams } from 'class-transformer';
 import {
   ArrayUnique,
   IsArray,
-  IsBoolean,
   IsNumber,
   IsOptional,
   IsString,
@@ -21,20 +20,13 @@ export class UpdateAccountDto {
   @Transform(({ value }: TransformFnParams) =>
     isString(value) ? value.trim() : value
   )
-  comment?: string;
+  comment?: string | null;
 
   @IsCurrencyCode()
   currency: string;
 
   @IsString()
   id: string;
-
-  /**
-   * @deprecated Use the "Exclude from Analysis" tag (`TAG_ID_EXCLUDE_FROM_ANALYSIS`) instead
-   */
-  @IsBoolean()
-  @IsOptional()
-  isExcluded?: boolean;
 
   @IsString()
   name: string;